+++ app/build.gradle
@@ -45,12 +45,14 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.1'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
     implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+
+
+    //Code verification view
+    implementation "com.github.fraggjkee:sms-confirmation-view:1.7.1"
 
     //view pager2
     implementation 'androidx.viewpager2:viewpager2:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
     //lottie
     def lottieVersion = "6.0.0"
@@ -59,8 +61,13 @@ dependencies {
     //Indicator
     implementation("com.tbuonomo:dotsindicator:4.3")
 
-    //Code verification view
-    implementation "com.github.fraggjkee:sms-confirmation-view:1.7.1"
+    //Retrofit2
+    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+
+    //OkHttpClient
+    implementation 'com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.2'
+    implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.2'
 
     implementation("androidx.recyclerview:recyclerview:1.3.0")
     // For control over item selection of both touch and mouse driven selection
