+++ app/src/main/java/com/abit8/geeksmentor/SplashActivity.kt
@@ -11,7 +11,7 @@ class SplashActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
 
         setContentView(R.layout.activity_splash)
-        Handler().postDelayed({            startActivity(Intent(this, MainActivity::class.java))
+        Handler().postDelayed({startActivity(Intent(this, MainActivity::class.java))
             finish()
         }, 3000)
     }
