+++ app/src/main/java/com/abit8/geeksmentor/base/BaseFragment.kt
@@ -1,0 +1,4 @@
+package com.abit8.geeksmentor.base
+
+class BaseFragment {
+}
