+++ app/src/main/java/com/abit8/geeksmentor/presentation/BestMentorListViewModel.kt
@@ -1,0 +1,7 @@
+package com.abit8.geeksmentor.presentation
+
+import androidx.lifecycle.ViewModel
+
+class BestMentorListViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
