+++ app/src/main/java/com/abit8/geeksmentor/presentation/MentorListFragment.kt
@@ -1,4 +1,4 @@
-package com.abit8.geeksmentor.ui.auth
+package com.abit8.geeksmentor.presentation
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.abit8.geeksmentor.R
 
-class SignInFragment : Fragment() {
+
+class MentorListFragment : Fragment() {
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_sign_in, container, false)
+        return inflater.inflate(R.layout.fragment_mentor_list, container, false)
     }
 }
