+++ app/src/main/java/com/abit8/geeksmentor/presentation/auth/LogInFragment.kt
@@ -1,0 +1,24 @@
+package com.abit8.geeksmentor.presentation.auth
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.abit8.geeksmentor.databinding.FragmentLogInBinding
+
+class LogInFragment : Fragment() {
+    private lateinit var binding: FragmentLogInBinding
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentLogInBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+}
