+++ app/src/main/java/com/abit8/geeksmentor/presentation/auth/PasswordForgetFragment2.kt
@@ -1,0 +1,18 @@
+package com.abit8.geeksmentor.presentation.auth
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.abit8.geeksmentor.R
+class PasswordForgetFragment2 : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_password_forget2, container, false)
+    }
+
+}
