+++ app/src/main/java/com/abit8/geeksmentor/presentation/home/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.abit8.geeksmentor.ui.profile
+package com.abit8.geeksmentor.presentation.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ProfileViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
