+++ app/src/main/java/com/abit8/geeksmentor/presentation/mentorprofile/EditMentorProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.abit8.geeksmentor.ui.auth
+package com.abit8.geeksmentor.presentation.mentorprofile
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.abit8.geeksmentor.R
 
-class LogInFragment : Fragment() {
+class EditMentorProfileFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_log_in, container, false)
+        return inflater.inflate(R.layout.fragment_edit_mentors_profile, container, false)
     }
 }
