+++ app/src/main/java/com/abit8/geeksmentor/presentation/mentorprofile/LanguageFragment.kt
@@ -1,0 +1,19 @@
+package com.abit8.geeksmentor.presentation.mentorprofile
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.abit8.geeksmentor.R
+
+class LanguageFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_language, container, false)
+    }
+}
