+++ app/src/main/java/com/abit8/geeksmentor/presentation/mentorprofile/MentorsNotificationFragment.kt
@@ -1,18 +1,19 @@
-package com.abit8.geeksmentor.ui.auth
+package com.abit8.geeksmentor.presentation.mentorprofile
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.abit8.geeksmentor.R
-class PasswordForgetFragment2 : Fragment() {
+
+class MentorsNotificationFragment : Fragment() {
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_password_forget2, container, false)
+        return inflater.inflate(R.layout.fragment_mentors_notification, container, false)
     }
-
 }
