+++ app/src/main/java/com/abit8/geeksmentor/presentation/profile/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.abit8.geeksmentor.ui.profile
+package com.abit8.geeksmentor.presentation.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
