+++ app/src/main/java/com/abit8/geeksmentor/presentation/profile/ProfileViewModel.kt
@@ -1,13 +1,13 @@
-package com.abit8.geeksmentor.ui.home
+package com.abit8.geeksmentor.presentation.profile
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class ProfileViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
