+++ app/src/main/java/com/abit8/geeksmentor/retrofit/LogInApi.kt
@@ -1,0 +1,8 @@
+package com.abit8.geeksmentor.retrofit
+
+
+import retrofit2.http.POST
+
+interface LogInApi {
+
+}
