+++ app/src/main/res/layout/fragment_best_mentor_list.xml
@@ -3,13 +3,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:id="@+id/edit_mentor_profile_info_fragment"
-    tools:context=".presentation.mentorprofile.EditMentorProfileFragment">
+    tools:context=".presentation.BestMentorListFragment">
 
-    <!-- TODO: Update blank fragment layout -->
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:text="Hello" />
 
 </FrameLayout>
