+++ app/src/main/res/layout/fragment_home.xml
@@ -10,7 +10,7 @@
         android:id="@+id/frame_home"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".ui.home.HomeFragment">
+        tools:context=".presentation.home.HomeFragment">
 
         <TextView
             android:layout_width="wrap_content"
@@ -136,8 +136,8 @@
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv1_main"
             android:layout_width="match_parent"
-            android:layout_height="145dp"
-            android:layout_marginTop="412dp"
+            android:layout_height="160dp"
+            android:layout_marginTop="400dp"
             tools:listitem="@layout/list1_item"
             android:orientation="horizontal"
             android:scrollbars="horizontal"
@@ -159,7 +159,7 @@
             style="@style/text_14"
             android:drawableRight="@drawable/line_right_black"
             android:layout_gravity="right"
-            android:layout_marginTop="592dp"
+            android:layout_marginTop="589dp"
             android:layout_marginRight="16dp"
             />
         <androidx.recyclerview.widget.RecyclerView
