+++ app/src/main/res/layout/fragment_language.xml
@@ -6,7 +6,7 @@
     android:layout_margin="@dimen/mrg_hor_16"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/mentors_profile_language_fragment"
-    tools:context=".ui.mentorprofile.LanguageFragment">
+    tools:context=".presentation.mentorprofile.LanguageFragment">
 
     <TextView
         android:id="@+id/btn_back"
@@ -23,7 +23,7 @@
         app:drawableLeftCompat="@drawable/left_arrow" />
 
     <RelativeLayout
-        android:id="@+id/btn_russian_language"
+        android:id="@+id/btn_mentors_profile_notification"
         android:layout_width="match_parent"
         android:layout_height="20dp"
         app:layout_constraintTop_toBottomOf="@+id/btn_back"
@@ -55,10 +55,10 @@
     </RelativeLayout>
 
     <RelativeLayout
-        android:id="@+id/btn_kyrgyz_language"
+        android:id="@+id/btn_mentors_profile_language"
         android:layout_width="match_parent"
         android:layout_height="20dp"
-        app:layout_constraintTop_toBottomOf="@+id/btn_russian_language"
+        app:layout_constraintTop_toBottomOf="@+id/btn_mentors_profile_notification"
         app:layout_constraintStart_toStartOf="parent"
         android:layout_marginTop="32dp">
 
@@ -87,11 +87,11 @@
     </RelativeLayout>
 
     <RelativeLayout
-        android:id="@+id/btn_english_language"
+        android:id="@+id/btn_mentors_profile_data"
         android:layout_width="match_parent"
         android:layout_marginTop="32dp"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/btn_kyrgyz_language"
+        app:layout_constraintTop_toBottomOf="@id/btn_mentors_profile_language"
         android:layout_height="20dp">
 
         <ImageView
