+++ app/src/main/res/layout/fragment_log_in.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.auth.LogInFragment"
+    tools:context=".presentation.auth.LogInFragment"
     android:background="@color/white_new">
     <TextView
         android:id="@+id/arrow_signup"
