+++ app/src/main/res/layout/fragment_mentor_profile.xml
@@ -6,7 +6,7 @@
     android:layout_margin="@dimen/mrg_hor_16"
     android:layout_height="match_parent"
     android:id="@+id/mentors_profile_fragment"
-    tools:context=".ui.mentorprofile.MentorProfileFragment">
+    tools:context=".presentation.mentorprofile.MentorProfileFragment">
 
     <de.hdodenhof.circleimageview.CircleImageView
         android:id="@+id/mentors_profile_photo"
@@ -57,7 +57,7 @@
         android:textSize="14sp"/>
 
     <RelativeLayout
-        android:id="@+id/btn_russian_language"
+        android:id="@+id/btn_mentors_profile_notification"
         android:layout_width="match_parent"
         android:layout_height="20dp"
         app:layout_constraintTop_toBottomOf="@+id/btn_edit_mentor_profile_info"
@@ -85,10 +85,10 @@
     </RelativeLayout>
 
     <RelativeLayout
-        android:id="@+id/btn_kyrgyz_language"
+        android:id="@+id/btn_mentors_profile_language"
         android:layout_width="match_parent"
         android:layout_height="20dp"
-        app:layout_constraintTop_toBottomOf="@+id/btn_russian_language"
+        app:layout_constraintTop_toBottomOf="@+id/btn_mentors_profile_notification"
         app:layout_constraintStart_toStartOf="parent"
         android:layout_marginTop="32dp">
 
@@ -112,11 +112,11 @@
     </RelativeLayout>
 
     <RelativeLayout
-        android:id="@+id/btn_english_language"
+        android:id="@+id/btn_mentors_profile_data"
         android:layout_width="match_parent"
         android:layout_marginTop="32dp"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/btn_kyrgyz_language"
+        app:layout_constraintTop_toBottomOf="@id/btn_mentors_profile_language"
         android:layout_height="20dp">
 
         <TextView
