+++ app/src/main/res/layout/fragment_mentors_data.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/mentors_profile_data_fragment"
-    tools:context=".ui.mentorprofile.MentorsDataFragment">
+    tools:context=".presentation.mentorprofile.MentorsDataFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
