+++ app/src/main/res/layout/fragment_mentors_notification.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/mentors_notification_fragment"
-    tools:context=".ui.mentorprofile.MentorsNotificationFragment">
+    tools:context=".presentation.mentorprofile.MentorsNotificationFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
