--- app/src/main/res/layout/fragment_on_board2.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:background="@color/black_new">
-
-    <ImageView
-        android:id="@+id/arrow_board2"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:src="@drawable/arrow_left_white"
-        android:layout_marginLeft="16dp"
-        android:layout_marginTop="78dp"
-        />
-
-    <ImageView
-        android:layout_width="340dp"
-        android:layout_height="340dp"
-        android:src="@drawable/shadow2"
-        android:layout_marginLeft="69dp"
-        android:layout_marginTop="200dp"/>
-    <ImageView
-        android:id="@+id/img_boarding_two"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/boarding_two"
-        android:layout_gravity="center|top"
-        android:layout_marginTop="103dp"
-        />
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/board2_like"
-        android:layout_marginLeft="95dp"
-        android:layout_marginTop="121dp"
-        />
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/only_best_mentors"
-        android:textColor="@color/white_new"
-        android:fontFamily="@font/roboto_medium"
-        android:textSize="26sp"
-        android:layout_marginLeft="16dp"
-        android:layout_marginTop="540dp"
-        />
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/btn_boarding2"
-        android:layout_width="328dp"
-        android:layout_height="58dp"
-        app:cornerRadius="10dp"
-        android:text="@string/continue_board1"
-        android:drawableRight="@drawable/arrow_right_black"
-        android:layout_gravity="bottom|center"
-        android:layout_marginBottom="84dp"
-        style="@style/btn_yellow_and_black"
-        />
-</FrameLayout>
