--- app/src/main/res/layout/fragment_on_board3.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:background="@color/black_new">
-    <TextView
-        android:id="@+id/arrow_board3"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:background="@drawable/arrow_left_white"
-        android:layout_marginTop="78dp"
-        android:layout_marginLeft="16dp"
-        />
-    <ImageView
-        android:layout_width="46dp"
-        android:layout_height="67dp"
-        android:src="@drawable/logo_yellow"
-        android:layout_gravity="center|top"
-        android:layout_marginTop="134dp"
-        />
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/enter_board3"
-        android:layout_width="328dp"
-        android:layout_height="58dp"
-        app:backgroundTint="@color/white_new"
-        app:cornerRadius="10dp"
-        android:text="@string/enter2"
-        android:textColor="@color/black_new"
-        android:textAllCaps="false"
-        android:fontFamily="@font/roboto_medium"
-        android:textSize="14sp"
-        android:layout_marginTop="249dp"
-        android:layout_gravity="top|center"
-        />
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/or"
-        android:textColor="@color/white_new"
-        android:fontFamily="@font/roboto_regular"
-        android:textSize="16dp"
-        android:layout_marginTop="321dp"
-        android:layout_gravity="center|top"
-        />
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/register_board3"
-        android:layout_width="328dp"
-        android:layout_height="58dp"
-        app:backgroundTint="@color/white_new"
-        app:cornerRadius="10dp"
-        android:text="@string/registration"
-        android:textColor="@color/black_new"
-        android:textAllCaps="false"
-        android:fontFamily="@font/roboto_medium"
-        android:textSize="14sp"
-        android:layout_marginTop="364dp"
-        android:layout_gravity="top|center"
-        />
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/btn_skip_board3"
-        android:layout_width="328dp"
-        android:layout_height="58dp"
-        app:cornerRadius="10dp"
-        android:text="@string/continue_board1"
-        android:drawableRight="@drawable/arrow_right_black"
-        android:layout_gravity="bottom|center"
-        android:layout_marginBottom="84dp"
-        style="@style/btn_yellow_and_black"
-        />
-</FrameLayout>
