+++ app/src/main/res/layout/fragment_password_forget.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.auth.PasswordForgetFragment"
+    tools:context=".presentation.auth.PasswordForgetFragment"
     android:background="@color/white_new">
     <TextView
         android:id="@+id/arrow_email"
