+++ app/src/main/res/layout/fragment_password_forget2.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.auth.PasswordForgetFragment2"
+    tools:context=".presentation.auth.PasswordForgetFragment2"
     >
     <TextView
         android:id="@+id/arrow_code"
