+++ app/src/main/res/layout/fragment_sign_in.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.auth.SignInFragment"
+    tools:context=".presentation.auth.SignInFragment"
     android:background="@color/white_new">
 
     <TextView
