+++ app/src/main/res/layout/list1_item.xml
@@ -2,7 +2,7 @@
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="123dp"
     android:layout_height="145dp"
-    android:layout_margin="16dp">
+    android:layout_margin="12dp">
     
     <com.google.android.material.imageview.ShapeableImageView
         android:id="@+id/rv1_img"
