+++ app/src/main/res/values/themes.xml
@@ -89,12 +89,13 @@
         <item name="android:fontFamily">@font/roboto_regular</item>
         <item name="textAllCaps">false</item>
         <item name="android:textStyle">bold</item>
+    </style>
 
-    <style name="btn_yellow_and_black">
+    <!--<style name="btn_yellow_and_black">
         <item name="android:textColor">#262523</item>
         <item name="android:backgroundTint">#FFE600</item>
         <item name="android:textSize">14sp</item>
         <item name="android:fontFamily">@font/roboto_medium</item>
         <item name="android:textAllCaps">false</item>
-    </style>
+    </style>-->
 </resources>
